plugins/gmail-imap/pyproject.toml
[project]
name = "coremind-plugin-gmail-imap"
version = "0.1.0"
description = "Gmail (IMAP IDLE) integration plugin for CoreMind."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Guillaume Gagnon" }]
requires-python = ">=3.12"
keywords = ["coremind", "plugin", "gmail", "imap"]

dependencies = [
    "aioimaplib>=1.1.0",
    "coremind>=0.0.1",
    "grpcio>=1.66.0",
    "structlog>=24.4.0",
]

[project.scripts]
coremind-plugin-gmail-imap = "coremind_plugin_gmail_imap.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coremind_plugin_gmail_imap"]
